MOdified as the Submission property changes
git-svn-id: https://coder.topcoder.com/tcs/clients/cronos/components@71409 fb370eea-3af6-4597-97f7-f7400a59c12a

+++ online_review_phases/trunk/src/java/main/com/cronos/onlinereview/phases/ReviewPhaseHandler.java
@@ -281,7 +281,7 @@ public class ReviewPhaseHandler extends AbstractPhaseHandler {
                 float aggScore = aggregations[iSub].getAggregatedScore();
 
                 // OrChange - Modified to update the submissions table instead of the resource_info table
-                submission.setIntialScore(Double.valueOf(String.valueOf(aggScore)).doubleValue());
+                submission.setIntialScore(Double.valueOf(String.valueOf(aggScore)));
                 getManagerHelper().getUploadManager().updateSubmission(submission, operator);
 
                 // Old Code Begins
